MERG COLISIUNILEEEE

+++ Components/Collisions/CollisionMatrix.hpp
@@ -13,8 +13,8 @@ class CollisionMatrix{
     size_t height;
     size_t nrSqrWidth;
     size_t nrSqrHeight;
-    size_t squareWidth;
-    size_t squareHeight;
+    float squareWidth;
+    float squareHeight;
     std::vector<Square> grid;
 public:
     CollisionMatrix(size_t width, size_t height, size_t nrSqrWidth, size_t nrSqrHeight);
